# docs/openapi/PNT2Y-OpenAPI3_PN-NationalRegistries.yaml
openapi: 3.0.0
info:
  title: PN-NationalRegistries REST APIs
  description: |-
    Nell’ambito del progetto Piattaforme Notifiche, il componente PN-NationalRegistries si occuperà di recuperare i dati dalle banche dati esterne previste. Tali banche dati esporranno delle interfacce web verso il componente PN-NationalRegistries, accessibili tramite piattaforma di interoperabilità.
  version: 0.11.5
servers:
  - url: http://127.0.0.1
...
paths:
...
  /national-registries-private/infocamere/legal:
    post:
      tags:
        - infoCamere
      summary: 'Questo servizio Il servizio consente di verificare se il codice fiscale della persona risulta legale rappresentante dell’impresa passata come parametro.'
      description: 'Questo servizio Il servizio consente di verificare se il codice fiscale della persona risulta legale rappresentante dell’impresa passata come parametro.'
      operationId: infoCamereLegal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InfoCamereLegalRequestBody'
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalOK'
        "400":
          description: Bad request
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalError'
        "401":
          description: Unauthorized
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalError'
        "404":
          description: Not Found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalError'
        "500":
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalError'
        "503":
          description: Service Unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalError'
  /national-registries-private/infocamere/legal-institutions:
    post:
      tags:
        - infoCamere
      summary: 'Il servizio restituisce la lista delle imprese che hanno come legale rappresentante la persona fisica indicata.'
      description: 'Il servizio restituisce la lista delle imprese che hanno come legale rappresentante la persona fisica indicata.'
      operationId: infoCamereLegalInstitutions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InfoCamereLegalInstitutionsRequestBody'
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsOK'
        "400":
          description: Bad request
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsError'
        "401":
          description: Unauthorized
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsError'
        "404":
          description: Not Found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsError'
        "500":
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsError'
        "503":
          description: Service Unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoCamereLegalInstitutionsError'
...
  /national-registries-private/inipec/digital-address:
    post:
      tags:
        - infoCamere
      summary: 'Consente di ottenere la PEC dell’impresa oppure del professionista corrispondente al codice fiscale al momento della consultazione. In caso di impresa si restituisce anche l’indirizzo della sede legale.'
      description: 'Consente di ottenere la PEC dell’impresa oppure del professionista corrispondente al codice fiscale al momento della consultazione. In caso di impresa si restituisce anche l’indirizzo della sede legale.'
      operationId: digitalAddressIniPEC
      parameters:
        - name: pn-national-registries-cx-id
          in: header
          schema:
            type: string
      requestBody:
        description: Consente di ottenere la PEC dell’impresa oppure del professionista corrispondente al codice fiscale al momento della consultazione. In caso di impresa si restituisce anche l’indirizzo della sede legale.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetDigitalAddressIniPECRequestBody'
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECOK'
        "400":
          description: Bad Request
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECError'
        "401":
          description: Unauthorized
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECError'
        "404":
          description: Not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECError'
        "500":
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECError'
        "503":
          description: Service Unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDigitalAddressIniPECError'
  /national-registries-private/registro-imprese/address:
    post:
      tags:
        - infoCamere
      summary: 'Consente di ottenere l’indirizzo della sede legale a cui corrisponde il codice fiscale al momento della consultazione'
      description: 'Consente di ottenere l’indirizzo della sede legale a cui corrisponde il codice fiscale al momento della consultazione'
      operationId: addressRegistroImprese
      requestBody:
        description: Consente di ottenere l’indirizzo della sede legale a cui corrisponde il codice fiscale al momento della consultazione
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAddressRegistroImpreseRequestBody'
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseOK'
        "400":
          description: Caso d'uso invalido
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseError'
        "401":
          description: Unauthorized
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseError'
        "404":
          description: Caso d'uso non trovato
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseError'
        "500":
          description: Internal server error
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseError'
        "503":
          description: Service Unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAddressRegistroImpreseError'
components:
  headers:
    RequestId:
      description: Identificativo univoco della richiesta
      schema:
        type: string
...
  schemas:
...
    GetDigitalAddressIniPECRequestBody:
      type: object
      properties:
        filter:
          type: object
          required:
            - taxId
            - correlationId
          properties:
            taxId:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
              example: AAAZZZ00H00T000Z
            correlationId:
              type: string
    GetDigitalAddressIniPECOK:
      type: object
      properties:
        correlationId:
          type: string
    GetDigitalAddressIniPECError:
      type: object
      properties:
        status:
          type: integer
          description: Contiene il codice HTTP dell'errore
        title:
          type: string
        detail:
          type: string
        traceId:
          type: string
        timestamp:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              element:
                type: string
              detail:
                type: string
    GetAddressRegistroImpreseRequestBody:
      type: object
      properties:
        filter:
          type: object
          required:
            - taxId
          properties:
            taxId:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
              example: AAAZZZ00H00T0
...
    GetAddressRegistroImpreseOK:
      type: object
      properties:
        dateTimeExtraction:
          type: string
          format: date-time
        taxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          format: password # NO EXTERNAL
          example: AAAZZZ00H00T000Z
        professionalAddress:
          type: object
          properties:
            description:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
            municipality:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
            province:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
            address:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
            zip:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
            stato:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
    GetAddressRegistroImpreseError:
      type: object
      properties:
        status:
          type: integer
          description: Contiene il codice HTTP dell'errore
        title:
          type: string
        detail:
          type: string
        traceId:
          type: string
        timestamp:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              element:
                type: string
              detail:
                type: string
              clientOperationId:
                type: string
...
    InfoCamereLegalInstitutionsRequestBody:
      type: object
      properties:
        filter:
          type: object
          required:
            - taxId
          properties:
            taxId:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
              example: AAAZZZ00H00T000Z
    InfoCamereLegalInstitutionsOK:
      type: object
      properties:
        dateTimeExtraction:
          type: string
        legalTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          format: password # NO EXTERNAL
          example: AAAZZZ00H00T000Z
        businessList:
          type: array
          items:
            $ref: '#/components/schemas/Business'
        code:
          type: string
        appName:
          type: string
        description:
          type: string
        timestamp:
          type: string
    Business:
      type: object
      properties:
        businessTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          format: password # NO EXTERNAL
        businessName:
          type: string
    InfoCamereLegalInstitutionsError:
      type: object
      properties:
        status:
          type: integer
          description: Contiene il codice HTTP dell'errore
        title:
          type: string
        detail:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              element:
                type: string
              detail:
                type: string
    InfoCamereLegalRequestBody:
      type: object
      properties:
        filter:
          type: object
          required:
            - taxId
            - vatNumber
          properties:
            taxId:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
              example: AAAZZZ00H00T000Z
            vatNumber:
              x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
              type: string
              format: password # NO EXTERNAL
              example: 12345678901
    InfoCamereLegalOK:
      type: object
      properties:
        dateTimeExtraction:
          type: string
          format: date-time
        taxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          format: password # NO EXTERNAL
          example: AAAZZZ00H00T000Z
        vatNumber:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          format: password # NO EXTERNAL
          example: BBBCCC00H00T000Z
        verificationResult:
          type: boolean
    InfoCamereLegalError:
      type: object
      properties:
        status:
          type: integer
          description: Contiene il codice HTTP dell'errore
        title:
          type: string
        detail:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              element:
                type: string
              detail:
                type: string
...